refactor: standalone manifests
