+++ angular.json
@@ -50,8 +50,7 @@
               "vendorChunk": true,
               "extractLicenses": false,
               "sourceMap": true,
-              "namedChunks": true,
-              "fileReplacements": []
+              "namedChunks": true
             },
             "production": {
               "fileReplacements": [
