+++ packages/components/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "components",
+  "version": "1.0.0"
+}
