+++ packages/components/plh/manifest.ts
@@ -1,0 +1,6 @@
+import { PLHComponentName } from ".";
+import { IComponentManifest } from "../types";
+
+export const PLH_COMPONENT_MANIFEST: IComponentManifest<PLHComponentName> = {
+  parent_point_box: { implicit: ["lottie_animation"] },
+};
