+++ src/app/shared/components/template/components/index.ts
@@ -61,9 +61,8 @@ import { TmplVideoComponent } from "./video";
 import { WorkshopsComponent } from "./layout/workshops_accordion";
 import { YoutubeComponent } from "./youtube/youtube.component";
 
-import { DEMO_COMPONENT_MAPPING } from "packages/components/demo";
-import { PLH_COMPONENT_MANIFEST, PLH_COMPONENT_MAPPING } from "packages/components/plh";
-import { IComponentManifest } from "packages/components/types";
+import { DEMO_COMPONENT_MAPPING } from "components/demo";
+import { PLH_COMPONENT_MAPPING } from "components/plh";
 
 /** All components should be exported as a single array for easy module import */
 export const TEMPLATE_COMPONENTS = [
@@ -197,26 +196,6 @@ const COMMON_COMPONENT_MAPPING = {
 
 /** Utility type of all named components from common mapping */
 export type ICommonComponentName = keyof typeof COMMON_COMPONENT_MAPPING;
-
-/** Manifest declaration of component dependencies (for use in optimisation) */
-const COMMON_COMPONENT_MANIFEST: IComponentManifest<ICommonComponentName> = {
-  carousel: { module: "SwiperModule" },
-  display_group: {
-    implicit: ["form", "advanced_dashed_box"],
-  },
-  lottie_animation: { module: "LottieModule" },
-  pdf: { assets: "/assets/comp-pdf", module: "NgxExtendedPdfViewerModule" },
-  slider: { module: "NouisliderModule" },
-};
-
-/**
- * List of components that are known to require specific changes to core build to function.
- * NOTE - this list is not comprehensive (should update as required)
- */
-export const TEMPLATE_COMPONENT_MANIFEST = {
-  ...COMMON_COMPONENT_MANIFEST,
-  ...PLH_COMPONENT_MANIFEST,
-};
 
 export const TEMPLATE_COMPONENT_MAPPING = {
   ...COMMON_COMPONENT_MAPPING,
